feat: taproot signature request (#278)
* compute taproot preimage not hash


* add script.mvoe


* taproot preimage for redeem tx


* update input index for sign to u32


* request sign to spend taproot output


* taproot


* Update bitcoin_lib/sources/sighash.move


* fmt


* fix test


* inline computation


* fix test


---------

+++ nBTC/tests/tx_composer_tests.move
@@ -44,7 +44,7 @@ fun create_unsign_redeem_tx_happy_case() {
     tx.inputs().length().do!(|i| {
         let sign_hash = create_segwit_preimage(
             &tx,
-            i, // input index
+            i as u32, // input index
             &script_code, // segwit nbtc spend key
             u64_to_le_bytes(utxos[i].value()), // amount
             0x01, // SIGNHASH_ALL
